requirements.txt
yfinance==0.2.43
pandas==2.2.2
numpy==1.26.4
pandas_ta==0.3.14b0
pyyaml==6.0.2
python-dateutil==2.9.0.post0
